+++ app/[locale]/admin/page.tsx
@@ -1,4 +1,4 @@
-import { getImportants, getSlides } from '@/apis/admin';
+import { getImportants, getSlides } from '@/apis/adminServer';
 
 import ImportantManagement from '@/components/admin/important/ImportantManagement';
 import SlideManagement from '@/components/admin/slide/SlideManagement';
@@ -24,14 +24,14 @@ export default async function AdminPage({ searchParams: { selected, page } }: Ad
   const pageNum = (page && parseInt(page)) || 1;
 
   if (selectedMenu === ADMIN_MENU.slide) {
-    const { posts, total } = await getSlides();
+    const { posts, total } = await getSlides(pageNum);
     return (
       <AdminPageLayout selectedMenu={selectedMenu}>
         {posts && <SlideManagement posts={posts} total={total} page={pageNum} />}
       </AdminPageLayout>
     );
   } else if (selectedMenu === ADMIN_MENU.important) {
-    const { posts, total } = await getImportants();
+    const { posts, total } = await getImportants(pageNum);
     return (
       <AdminPageLayout selectedMenu={selectedMenu}>
         {posts && <ImportantManagement posts={posts} total={total} page={pageNum} />}
