+++ components/admin/important/ImportantManagement.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from 'next/navigation';
 import { useReducer } from 'react';
 
-import { batchUnimportant } from '@/actions/adminActions';
+import { batchUnimportantAction } from '@/actions/adminActions';
 
 import { StraightNode } from '@/components/common/Nodes';
 import Pagination from '@/components/common/Pagination';
@@ -72,14 +72,9 @@ export default function ImportantManagement({ posts, page, total }: ImportantMan
   };
 
   const handleBatchUnimportant = async () => {
-    const result = await batchUnimportant(selectedPostIdentifiers);
+    const result = await batchUnimportantAction(selectedPostIdentifiers);
     if (result?.error) {
       errorToast('중요 안내를 해제하지 못했습니다.');
-      if (result.error instanceof Error) {
-        console.error(result.error.message);
-      } else {
-        throw result.error;
-      }
     } else {
       successToast('중요 안내를 해제했습니다.');
       resetSelectedPosts();
