+++ components/admin/slide/SlideManagement.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from 'next/navigation';
 import { useReducer } from 'react';
 
-import { batchUnslide } from '@/actions/adminActions';
+import { batchUnslideAction } from '@/actions/adminActions';
 
 import { StraightNode } from '@/components/common/Nodes';
 import Pagination from '@/components/common/Pagination';
@@ -65,14 +65,9 @@ export default function SlideManagement({ posts, page, total }: SlideManagementP
   };
 
   const handleBatchUnslide = async () => {
-    const result = await batchUnslide(selectedPostIds);
+    const result = await batchUnslideAction(selectedPostIds);
     if (result?.error) {
       errorToast('슬라이드를 해제하지 못했습니다.');
-      if (result.error instanceof Error) {
-        console.error(result.error.message);
-      } else {
-        throw result.error;
-      }
     } else {
       successToast('슬라이드를 해제했습니다.');
       resetSelectedPosts();
@@ -97,11 +92,11 @@ export default function SlideManagement({ posts, page, total }: SlideManagementP
       />
       <BatchAction
         selectedCount={selectedPostIds.size}
-        buttonText="일괄 슬라이드쇼 해제"
+        buttonText="일괄 슬라이드 해제"
         onClickButton={() =>
           openModal(
             <AlertModal
-              message="정말 선택된 슬라이드쇼를 모두 해제하시겠습니까?"
+              message="정말 선택된 슬라이드를 모두 해제하시겠습니까?"
               confirmText="해제"
               onConfirm={handleBatchUnslide}
             />,
